requirements.txt
faiss-cpu
sentence-transformers
flask
requests
python-dotenv
